Restructure library to extract the hcaptcha builder structure to seperate module and re-export the necessary public types from the library only.
